fix: update check-prompts output for CI compatibility and refactor report module
- Update `check-prompts` CLI output in plain mode to support CI parsing.
- Add 'Refactored Suggestions:' header and explicit failure message.
- Refactor `src/agent_scorecard/report.py` to inline logic and remove dependency on `report_sections.py`.
- Add unit tests for prompt output formatting.
